Struct poseidon::poseidon::Poseidon

pub struct Poseidon { /* private fields */ }

Defines the storage of your contract. Add new fields to the below struct in order to add new static storage fields to your contract.